Chat

+++ backend/marathon/src/main/java/com/kmozdzen/marathon/service/yourplanService/YourPlanServiceImpl.java
@@ -8,8 +8,11 @@ import com.kmozdzen.marathon.response.AnswersResponse;
 import com.kmozdzen.marathon.respository.RunRepository;
 import com.kmozdzen.marathon.respository.UserRepository;
 import com.kmozdzen.marathon.respository.YourPlanRepository;
+import com.kmozdzen.marathon.service.answerService.AnswerService;
+import com.kmozdzen.marathon.service.userService.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
@@ -20,12 +23,14 @@ public class YourPlanServiceImpl implements YourPlanService{
     private YourPlanRepository yourPlanRepository;
     private UserRepository userRepository;
     private RunRepository runRepository;
+    private AnswerService answerService;
 
     @Autowired
-    public YourPlanServiceImpl(YourPlanRepository yourPlanRepository, UserRepository userRepository, RunRepository runRepository) {
+    public YourPlanServiceImpl(YourPlanRepository yourPlanRepository, UserRepository userRepository, RunRepository runRepository, AnswerService answerService) {
         this.yourPlanRepository = yourPlanRepository;
         this.userRepository = userRepository;
         this.runRepository = runRepository;
+        this.answerService = answerService;
     }
 
     @Override
@@ -34,21 +39,20 @@ public class YourPlanServiceImpl implements YourPlanService{
     }
 
     @Override
-    public YourPlan create(String email, AnswersResponse answersResponse, LocalDate date) {
+    public YourPlan create(String email, AnswersResponse answersResponse, LocalDate date, LocalTime mmTime) {
         System.out.println(answersResponse);
 
         Galloway galloway = new Galloway();
 
-        LocalTime raceTime = LocalTime.of(0,10,0);
-        galloway.create(answersResponse, date, raceTime);
-
+        galloway.create(answersResponse, date, mmTime);
 
         if(yourPlanRepository.findByUserEmail(email) == null){
             User user = userRepository.findByEmail(email);
 
             YourPlan yourPlan = new YourPlan();
             yourPlan.setRaceDate(date);
-            yourPlan.setName("default");
+            yourPlan.setName(answersResponse.getRaceName());
+            yourPlan.setMmTime(mmTime);
             yourPlan.setUser(user);
 
             yourPlanRepository.save(yourPlan);
@@ -89,6 +93,13 @@ public class YourPlanServiceImpl implements YourPlanService{
                     run.setTime(time);
                 }
 
+                if(runDate.getValue().get("additionalInfo") != null){
+                    String additionalInfo = (String) runDate.getValue().get("additionalInfo");
+                    if(additionalInfo != "")
+                        run.setAdditionalInfo(additionalInfo);
+                }
+
+                run.setRunCheck(false);
                 run.setYourPlan(yourPlan);
                 runRepository.save(run);
             }
@@ -99,9 +110,15 @@ public class YourPlanServiceImpl implements YourPlanService{
 
     @Override
     public void remove(int id) {
-        YourPlan yourPlan = yourPlanRepository.findById(id).orElse(null);
-        if (yourPlan != null) {
-            yourPlanRepository.delete(yourPlan);
+        User user = userRepository.findById(id).orElse(null);
+
+        if(user != null){
+            YourPlan yourPlan = yourPlanRepository.findById(user.getYourPlan().getIdYourPlan()).orElse(null);
+            if (yourPlan != null) {
+                answerService.clearUserAnswers(user);
+                yourPlanRepository.delete(yourPlan);
+            }
         }
     }
+
 }
